Create folder for useInputMask
I decided to create a folder so I can store the logic for the masks (currency,
date, ..) in another file. Maybe I could store them in a 'utils' folder but,
since these functions will only be used in one place I think it is better if
they are closer to where they are used.